+++ src/components/modal-overlay/modal-overlay.module.css
@@ -2,6 +2,8 @@
   width: 100vw;
   height: 100vh;
   position: fixed;
+  top: 0;
+  left: 0;
   background-color: rgba(0, 0, 0, 0.5);
   background-size: contain;
   display: flex;
